-- minor fixes to remove some warning messages generated by gcc -Wall (trouble-shooting a problem on octagon) (--> b0.9.5)
git-svn-id: https://codeforge.lbl.gov/anonscm/fastbit/trunk@49 9ec3b2dc-5744-0410-9993-b8723341f1db

+++ src/irange.cpp
@@ -221,7 +221,7 @@ int ibis::range::read(const char* f) {
 	minval.swap(dbl);
     }
     ierr = UnixSeek(fdes, end, SEEK_SET);
-    if (ierr != end) {
+    if (ierr != static_cast<int>(end)) {
 	UnixClose(fdes);
 	remove(fnm.c_str());
 	LOGGER(ibis::gVerbose >= 1)
@@ -358,7 +358,7 @@ int ibis::range::read(int fdes, uint32_t start, const char *fn) {
 	minval.swap(dbl);
     }
     ierr = UnixSeek(fdes, end, SEEK_SET);
-    if (ierr != end) {
+    if (ierr != static_cast<int>(end)) {
 	UnixClose(fdes);
 	clear();
 	return -4;
